Create tasks for Antimatter project
This commit reads a folder with the images, creates a task per viewid. The
user has to pass a url where the images will be accessible.

The URL lacks the zoom level and the extension, so the script will create
for the task an info field like this:

url = 'http://domain/image_viewid_ifr

The template will have to fill in the zoom and .jpg extension